requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
google-auth
google-auth-oauthlib
google-api-python-client
